backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.7
redis==5.0.1
PyJWT==2.8.0
celery==5.3.4
Pillow==10.1.0
python-decouple==3.8
django-extensions==3.2.3
drf-spectacular==0.27.0
qrcode[pil]==7.4.2
sendgrid==6.10.0
python-magic-bin==0.4.14
opencv-python==4.13.0.90
numpy==2.3.4
pandas==2.0.3
scikit-learn==1.3.2
transformers==4.35.2
torch==2.1.1
requests==2.31.0
cryptography==41.0.7
bcrypt==4.1.1
django-phonenumber-field==7.3.0
phonenumbers==8.13.25
django-storages==1.14.2
boto3==1.34.0
django-allauth==0.57.0
dj-rest-auth==5.0.2
django-oauth-toolkit==1.7.1
django-redis==5.4.0
gunicorn==21.2.0
whitenoise==6.6.0
mysqlclient==2.2.0
